Have thread meta data store use remote identity for matching
